Edicion
ombe

+++ backend/routes/movieRoutes.js
@@ -7,6 +7,9 @@ express.Router();
 const Movie =
 require("../models/Movie");
 
+const authMiddleware =
+require("../middleware/authMiddleware");
+
 router.get(
 
     "/",
@@ -95,16 +98,34 @@ router.post(
 
     "/",
 
+    authMiddleware,
+
     async (req, res) => {
 
-        const movie =
-        await Movie.create(
-            req.body
-        );
+        try {
 
-        res.status(201).json(
-            movie
-        );
+            const movie =
+            await Movie.create({
+
+                ...req.body,
+
+                creadoPor:
+                req.user.userId
+
+            });
+
+            res.status(201).json(
+                movie
+            );
+
+        } catch (error) {
+
+            res.status(500).json({
+
+                message:error.message
+
+            });
+        }
     }
 );
 
@@ -112,9 +133,34 @@ router.put(
 
     "/:id",
 
+    authMiddleware,
+
     async (req, res) => {
 
         try {
+
+            const movie =
+            await Movie.findById(
+                req.params.id
+            );
+
+            if (
+
+                movie.creadoPor.toString()
+
+                !==
+
+                req.user.userId
+
+            ) {
+
+                return res.status(403).json({
+
+                    message:
+                    "No autorizado"
+
+                });
+            }
 
             const updatedMovie =
 
@@ -137,7 +183,9 @@ router.put(
         } catch (error) {
 
             res.status(500).json({
+
                 message:error.message
+
             });
         }
     }
@@ -147,17 +195,55 @@ router.delete(
 
     "/:id",
 
+    authMiddleware,
+
     async (req, res) => {
 
-        await Movie.findByIdAndDelete(
-            req.params.id
-        );
+        try {
 
-        res.json({
-            message:"Película eliminada"
-        });
+            const movie =
+            await Movie.findById(
+                req.params.id
+            );
+
+            if (
+
+                movie.creadoPor.toString()
+
+                !==
+
+                req.user.userId
+
+            ) {
+
+                return res.status(403).json({
+
+                    message:
+                    "No autorizado"
+
+                });
+            }
+
+            await Movie.findByIdAndDelete(
+                req.params.id
+            );
+
+            res.json({
+
+                message:
+                "Película eliminada"
+
+            });
+
+        } catch (error) {
+
+            res.status(500).json({
+
+                message:error.message
+
+            });
+        }
     }
 );
-
 module.exports =
 router;
